refactor(sentry): split reporting from the error boundary
Separate the generic ErrorBoundary (catch + fallback UI + onError callback)
from SentryErrorBoundary (a thin wrapper that injects Sentry reporting).

Fixes a real robustness gap: the Sentry capture call previously ran
unguarded in componentDidCatch, so if it threw, the error escaped the
boundary and the fallback UI never rendered. ErrorBoundary now invokes
onError inside try/catch, so a reporting failure can never prevent the
fallback from showing. The generic boundary is also reusable/testable
without any Sentry dependency.
