+++ redisinsight/ui/src/components/error-boundary/index.ts
@@ -1,1 +1,2 @@
+export { default as ErrorBoundary } from './ErrorBoundary'
 export { default as SentryErrorBoundary } from './SentryErrorBoundary'
